DataDrivenEnzymeRateEquations

Documentation for DataDrivenEnzymeRateEquations.

DataDrivenEnzymeRateEquations.calculate_all_parameter_removal_codes

DataDrivenEnzymeRateEquations.fit_rate_equation

DataDrivenEnzymeRateEquations.forward_selection_next_param_removal_codes

DataDrivenEnzymeRateEquations.loss_rate_equation

DataDrivenEnzymeRateEquations.param_rescaling

DataDrivenEnzymeRateEquations.param_subset_select

DataDrivenEnzymeRateEquations.reverse_selection_next_param_removal_codes

DataDrivenEnzymeRateEquations.test_rate_equation

DataDrivenEnzymeRateEquations.calculate_all_parameter_removal_codes — Method

Generate all possibles codes for ways that mirror params for a and i states of MWC enzyme can be removed from the rate equation

DataDrivenEnzymeRateEquations.fit_rate_equation — Method

fit_rate_equation(
    rate_equation::Function,
    data::DataFrame,
    metab_names::Tuple,
    param_names::Tuple;
    n_iter = 20

)

Fit rate_equation to data and return loss and best fit parameters.

Arguments

rate_equation::Function: Function that takes a NamedTuple of metabolite concentrations (with metab_names keys) and parameters (with param_names keys) and returns an enzyme rate.

data::DataFrame: DataFrame containing the data with column Rate and columns for each metab_names where each row is one measurement. It also needs to have a column source that contains a string that identifies the source of the data. This is used to calculate the weights for each figure in the publication.

metab_names::Tuple: Tuple of metabolite names that correspond to the metabolites of rate_equation and column names in data.

param_names::Tuple: Tuple of parameter names that correspond to the parameters of rate_equation.

n_iter::Int: Number of iterations to run the fitting process.

Returns

loss::Float64: Loss of the best fit.

params::NamedTuple: Best fit parameters with param_names keys

Example

using DataFrames
data = DataFrame(
    Rate = [1.0, 2.0, 3.0],
    A = [1.0, 2.0, 3.0],
    source = ["Figure 1", "Figure 1", "Figure 2"]
)
rate_equation(metabs, params) = params.Vmax * metabs.S / (1 + metabs.S / params.K_S)
fit_rate_equation(rate_equation, data, (:A,), (:Vmax, :K_S))

DataDrivenEnzymeRateEquations.forward_selection_next_param_removal_codes — Method

Calculate nt_param_removal_codes with num_params including non-zero term combinations for codes (excluding alpha terms) in each previous_param_removal_codes that has num_params-1

DataDrivenEnzymeRateEquations.loss_rate_equation — Method

Loss function used for fitting that calculate log of ratio of rate equation predicting of rate and rate data

DataDrivenEnzymeRateEquations.param_rescaling — Method

Rescaling of fitting parameters from [0, 10] scale that optimizer uses to actual values

DataDrivenEnzymeRateEquations.param_subset_select — Method

Function to convert parameter vector to vector where some params are equal to 0, Inf or each other based on ntparamremoval_code

DataDrivenEnzymeRateEquations.reverse_selection_next_param_removal_codes — Method

Calculate param_removal_codes with num_params including zero term combinations for codes (excluding alpha terms) in each previous_param_removal_codes that has num_params+1

DataDrivenEnzymeRateEquations.test_rate_equation — Method

Function to calculate loss for a given rate_equation and nt_fitted_params on data that was not used for training