DataDrivenEnzymeRateEqs

Documentation for DataDrivenEnzymeRateEqs.

DataDrivenEnzymeRateEqs.calculate_all_parameter_removal_codes

DataDrivenEnzymeRateEqs.data_driven_rate_equation_selection

DataDrivenEnzymeRateEqs.fit_rate_equation

DataDrivenEnzymeRateEqs.forward_selection_next_param_removal_codes

DataDrivenEnzymeRateEqs.generate_metab_names

DataDrivenEnzymeRateEqs.generate_param_names

DataDrivenEnzymeRateEqs.loss_rate_equation

DataDrivenEnzymeRateEqs.param_rescaling

DataDrivenEnzymeRateEqs.param_subset_select

DataDrivenEnzymeRateEqs.reverse_selection_next_param_removal_codes

DataDrivenEnzymeRateEqs.test_rate_equation

DataDrivenEnzymeRateEqs.@derive_general_mwc_rate_eq

DataDrivenEnzymeRateEqs.calculate_all_parameter_removal_codes — Method

Generate all possibles codes for ways that mirror params for a and i states of MWC enzyme can be removed from the rate equation

DataDrivenEnzymeRateEqs.data_driven_rate_equation_selection — Method

data_driven_rate_equation_selection(
    general_rate_equation::Function,
    data::DataFrame,
    metab_names::Tuple{Symbol,Vararg{Symbol}},
    param_names::Tuple{Symbol,Vararg{Symbol}},
    range_number_params::Tuple{Int,Int},
    forward_model_selection::Bool,
)

This function is used to perform data-driven rate equation selection using a general rate equation and data. The function will select the best rate equation by iteratively removing parameters from the general rate equation and finding an equation that yield best test scores on data not used for fitting.

Arguments

general_rate_equation::Function: Function that takes a NamedTuple of metabolite concentrations (with metab_names keys) and parameters (with param_names keys) and returns an enzyme rate.

data::DataFrame: DataFrame containing the data with column Rate and columns for each metab_names where each row is one measurement. It also needs to have a column source that contains a string that identifies the source of the data. This is used to calculate the weights for each figure in the publication.

metab_names::Tuple: Tuple of metabolite names that correspond to the metabolites of rate_equation and column names in data.

param_names::Tuple: Tuple of parameter names that correspond to the parameters of rate_equation.

range_number_params::Tuple{Int,Int}: A tuple of integers representing the range of the number of parameters of generalrateequation to search over.

forward_model_selection::Bool: A boolean indicating whether to use forward model selection (true) or reverse model selection (false).

Returns nothing, but saves a csv file for each num_params with the results of the training for each combination of parameters tested and a csv file with test results for top 10% of the best results with each number of parameters tested.

DataDrivenEnzymeRateEqs.fit_rate_equation — Method

fit_rate_equation(
    rate_equation::Function,
    data::DataFrame,
    metab_names::Tuple,
    param_names::Tuple;
    n_iter = 20

)

Fit rate_equation to data and return loss and best fit parameters.

Arguments

rate_equation::Function: Function that takes a NamedTuple of metabolite concentrations (with metab_names keys) and parameters (with param_names keys) and returns an enzyme rate.

data::DataFrame: DataFrame containing the data with column Rate and columns for each metab_names where each row is one measurement. It also needs to have a column source that contains a string that identifies the source of the data. This is used to calculate the weights for each figure in the publication.

metab_names::Tuple: Tuple of metabolite names that correspond to the metabolites of rate_equation and column names in data.

param_names::Tuple: Tuple of parameter names that correspond to the parameters of rate_equation.

n_iter::Int: Number of iterations to run the fitting process.

Returns

loss::Float64: Loss of the best fit.

params::NamedTuple: Best fit parameters with param_names keys

Example

using DataFrames
data = DataFrame(
    Rate = [1.0, 2.0, 3.0],
    A = [1.0, 2.0, 3.0],
    source = ["Figure 1", "Figure 1", "Figure 2"]
)
rate_equation(metabs, params) = params.Vmax * metabs.S / (1 + metabs.S / params.K_S)
fit_rate_equation(rate_equation, data, (:A,), (:Vmax, :K_S))

DataDrivenEnzymeRateEqs.forward_selection_next_param_removal_codes — Method

Calculate nt_param_removal_codes with num_params including non-zero term combinations for codes (excluding alpha terms) in each previous_param_removal_codes that has num_params-1

DataDrivenEnzymeRateEqs.generate_metab_names — Method

Generate the names of the metabolites for the rate equation using the same input as @derivegeneralmwcrateeq

DataDrivenEnzymeRateEqs.generate_param_names — Method

Generate the names of the parameters for the rate equation using the same input as @derivegeneralmwcrateeq

DataDrivenEnzymeRateEqs.loss_rate_equation — Method

Loss function used for fitting that calculate log of ratio of rate equation predicting of rate and rate data

DataDrivenEnzymeRateEqs.param_rescaling — Method

Rescaling of fitting parameters from [0, 10] scale that optimizer uses to actual values

DataDrivenEnzymeRateEqs.param_subset_select — Method

Function to convert parameter vector to vector where some params are equal to 0, Inf or each other based on ntparamremoval_code

DataDrivenEnzymeRateEqs.reverse_selection_next_param_removal_codes — Method

Calculate param_removal_codes with num_params including zero term combinations for codes (excluding alpha terms) in each previous_param_removal_codes that has num_params+1

DataDrivenEnzymeRateEqs.test_rate_equation — Method

Function to calculate loss for a given rate_equation and nt_fitted_params on data that was not used for training

DataDrivenEnzymeRateEqs.@derive_general_mwc_rate_eq — Macro

derive_general_mwc_rate_eq(metabs_and_regulators_kwargs...)

Derive a function that calculates the rate of a reaction using the general MWC rate equation given the list of substrates, products, and regulators that bind to specific cat or reg sites.

The general MWC rate equation is given by:

\[Rate = \frac{{V_{max}^a \prod_{i=1}^{n} \left(\frac{S_i}{K_{a, i}}\right) - V_{max}^a_{rev} \prod_{i=1}^{n} \left(\frac{P_i}{K_{a, i}}\right) \cdot Z_{a, cat}^{n-1} \cdot Z_{a, reg}^n + L \left(V_{max}^i \prod_{i=1}^{n} \left(\frac{S_i}{K_{i, i}}\right) - V_{max}^i_{rev} \prod_{i=1}^{n} \left(\frac{P_i}{K_{i, i}}\right)\right) \cdot Z_{i, cat}^{n-1} \cdot Z_{i, reg}^n}}{Z_{a, cat}^n \cdot Z_{a, reg}^n + L \cdot Z_{i, cat}^n \cdot Z_{i, reg}^n}\]

where:

$V_{max}^a$ is the maximum rate of the forward reaction

$V_{max}^a_{rev}$ is the maximum rate of the reverse reaction

$V_{max}^i$ is the maximum rate of the forward reaction

$V_{max}^i_{rev}$ is the maximum rate of the reverse reaction

$S_i$ is the concentration of the $i^{th}$ substrate

$P_i$ is the concentration of the $i^{th}$ product

$K_{a, i}$ is the Michaelis constant for the $i^{th}$ substrate

$K_{i, i}$ is the Michaelis constant for the $i^{th}$ product

$Z_{a, cat}$ is the allosteric factor for the catalytic site

$Z_{i, cat}$ is the allosteric factor for the catalytic site

$Z_{a, reg}$ is the allosteric factor for the regulatory site

$Z_{i, reg}$ is the allosteric factor for the regulatory site

$L$ is the concentration of the ligand

$n$ is the oligomeric state of the enzyme

Arguments

metabs_and_regulators_kwargs...: keyword arguments that specify the substrates, products, catalytic sites, regulatory sites, and other parameters of the reaction.

Returns

A function that calculates the rate of the reaction using the general MWC rate equation

A tuple of the names of the metabolites and parameters used in the rate equation